Class Match is now public
